Station Spawning, Tracking UI and Assignment Management
Fixed the station spawning, assembly and data handling. You can create the data for the prefabs and the editor and they spawn in well. The order generation is working well, and they carry over to the main game screen when accepted. From there you can now "track" the orders, which will display an icon on the hud on where that location is.